server: handle switching repo in the protocol implementation
This is a bit of a trade-off, it moves more logic into the
protocol that we may actually want to replace at some point.
However, that's not very likely to happen soon.

On the other hand, this makes GitServerBackend almost the
same as LocalRepo, so we can merge these things next.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>