server: move index sending into GitServerBackend
Instead of requiring the backend to provide the full index as
e.g. an mmap object or such, move the sending of it down into
the backend, with a send_size() function argument to deal with
that part of the protocol.

This means we can stream the index in the future, e.g. if the
repo is encrypted.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>