init/repo/server: make init go through repo abstraction
This is not ideal, I'd like to be able to just init the
repo and get one back after the create, but that causes
a lot of trouble with the commands calling git repo
functions directly, so do it with a classmethod for now.

Also expand this to the server, further reducing the
GitServerBackend class code.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>