git: change PackWriter() objcache_maker so None is default
Change the way we use objcache_maker and don't use None as
the function when none is desired, but rather use a lambda
that returns None, and then let =None mean the default.

This simplifies things as you can pass =None to have the
default even a few levels down in functions, right now it
simplifies the server, but the next patch will push that
into the Repo.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>