server: remove AbstractServerBackend/GitServerBackend
Now that we've pushed everything down into LocalRepo, we
can finally get rid of the server backend and just use the
LocalRepo class directly.

Due to the previous patch with rev-list, we cannot actually
have a RemoteRepo used by the server, but that's OK, we
don't really have a use case for that.

I haven't quite figured out the semantics of check_repo_or_die
and so this makes a tiny new wrapper class to call it.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>