bup: support encrypted repositories
Now, with a lot of pieces in place, add support for encrypted
repositories. The code contains more detailed documentation on
the actual encryption design.

An encrypted repository is a config repository, and the newly
added 'bup genkey' command can be used to create keys (and a
partial template for the configuration file.)

Also add documentation for all this, including for the storage
abstraction and file storage backend (here since it's previously
not actually useful and refers to the encrypted repository.)

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>