bup: support (optional) tree splitting
As hashed out previously on the mailing list, for some workloads
the current bup behaviour wrt. trees is rather suboptimal as the
tree objects can get very large. In my pathological case (backing
up Maildir email storage), I end up backing up a 36MB+ tree every
single time a backup is made, and the backup in the end ends up
with 40-60% (depending on the day) of the pack size just for all
those tree objects.

Add the ability to store split trees to avoid this situation.

The way it works is that - building on the previous patch to have
a record-based hashsplitter - a tree is now no longer necessarily
stored as-is, but subjected to hashsplitting. Since it should be
a tree object as far as git is concerned even then, split only on
record boundaries and make a new tree structure like this:

  folder/{.bupm,aa,bb,...,zz}

becomes

  folder/1.bupd
  folder/.b/{.bupm,aa,...,ii}
  folder/j/{jj,...,rr}
  folder/s/{ss,...,zz}

Note how the first filename of a sub-tree is abbreviated to its
shortest valid unique substring and then used as its name to later
allow faster lookups without reconstructing everything; this isn't
yet implemented, however.

Note also that this is done recursively, and the "<level>.bupd" is
only added at the top-most level, this level anyway needs to be
changed every time anything in this folder is touched.

The choice to store an empty sentinel file whose name indicates
the recursion level needed to find the real files was mostly due
to the fact that all bup internal functions for tree reading pass
around only three tree oid or contents, and not the (un-demangled)
name of the object. This made other schemes I considered far more
difficult to implement.

Note that the .bupm files for metadata are stored at the leaves,
except the one for the root directory. This causes a rather small
object to be stored for the root (typically, at least), but this
can't be avoided.

Finally, note that the input to the record hashsplitting isn't the
full tree object content, but only the (unmangled) filename (even
before abbreviating to have more data); this ensures that changes
to the content of any files don't have ripple effects on the tree
splitting, which could otherwise create a lot of new tree objects
only because a file changed content.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
---
v2:
 * rebase to not require all my other changes
 * implement abbreviation
 * make this optional
 * move reading code into vfs

v3:
 * move into own treesplit.py file, it cannot be in hashsplit as
   that is imported by git.py so can't import git.py
 * add notes about this to the DESIGN notes
 * add note to the bup-save man page

v4:
 * fix bug in name abbreviation and add tests for it

v5:
 * use bytes in the treesplit tests, sorry

v6:
 * more b'' usage
 * don't put the sentinel file first, git assumes tree objects are
   sorted for searching in them
 * add a test for this (git ls-tree being able to look into a split
   tree object)

v7:
 * use correct key=git.shalist_item_sort_key for sorting

v8:
 * don't call "git config" over and over again, just get the value
   once at startup

v9:
 * use opttype='bool'

v10:
 * apply hashsplitting only to <mode string> <filename>

v11:
 * apply hashsplitting only to <filename> to avoid ripple effects
   if a file changes content and is then split, or similar
 * rebase on top of the refactorings to be able to ...
 * store .bupm in the leaves, in part to ...
 * make the vfs not have to restore everything in one chunk

v12:
 * improve test by overriding the RecordHashSplitter to force
   two levels
 * improve test to check there's no .bupm in intermediate levels
 * improve code to actually not write .bupm there

v13:
 * use unabbreviated names for further hashsplitting to actually
   get multiple levels, given the previous changes in v11 that
   avoid ripple effects due to mode (or mode because of content,
   given file splitting) changes

v14:
 * fix tree reading when metadata isn't requested, this only
   surfaced after I optimised 'bup ls' to no longer need it.