git: PackIdxList/PackIdx: add the ability to return the offset
Add the ability to not just query _whether_ an object exists, and
if so return in which pack (or midx), but also to return the offset
at which it exists.

If want_offs is True, we cannot directly return the data from the
midx since it doesn't know about the offset, in this case look up
the object again in the original pack idx.

If instead we were to not use the midx, large repositories with a
lot of pack files quickly run out of file descriptors/mmap space,
and using the midx to know where to look is also more efficient.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>