storage: add storage abstraction
This will be used by the encrypted repo - we may want to store
that in arbitrary storage, e.g. cloud storage, and we don't
really need a lot of API to access it.

Supports the following:
 * writing a file - you can stream it out, there's no way to
   seek back to the beginning and modify a file right now
 * reading a file - this needs to be able to seek in order
   to read arbitrary pieces of the data
 * listing the files - need to generate a list of idx files
   to synchronize a repo back if the cache was removed
 * ability to atomically overwrite a file with a previously
   read one (failing on races)

We may want to extend this in the future to let even real git
repositories to go through the abstraction, but for now this
isn't supported and this will be used only by the encrypted
repo where you can't reasonably mmap() a file anyway since it
has to be decrypted.

Also add a file-based backend here, with some validation on
the APIs.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>