repo: support config-file based repositories
In order to support encrypted repositories, extend the
"remote" repository syntax by a URL-type config:// that
encodes the path to the configuration file and the name
of the repository in the configuration.

To specify such a repository, use

	config:///path/to/file.conf

in which case the named configuration file will be loaded.

In this configuration file, the bup.type setting must be
present to indicate the type of the repository. Further
access configuration then depends on the repo type.

Note that configuration values such as bup.split-trees
that aren't needed to access the repo, but rather are
properties of the repo, are stored in the repo itself.
They can be read and set with the bup config command.

In this patch, no such repo type is supported, the next
patch adds the 'Encrypted' type.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>