bup-get: add ability to split data & metadata
If bup get is writing into an encrypted repository that has
bup.splitmeta=true, then it's useful if it can separate out
between data and metadata. Teach bup get to do that, which
requires extending the just_write() API with an indication
of data/metadata.

FIXME: handle split .bupm correctly!

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>