repo: honour git compression level configuration
Make bup honour the git compression level, using, in
order of precedence:
 * level specified on command line (passed to the
   PackWriter instantiation)
 * git config pack.compression
 * git config core.compression
 * default to the previous default of 1

Note that the git config values are taken from the *repository*
configuration, so they come from the remote repository when the
data is stored remotely. This differs from the handling of the
pack.packSizeLimit option, which currently comes only from the
*local* bup repo (typically ~/.bup/) even when a backup is made
to a remote repository.

That last bullet makes it compatible with old versions of
bup, whereas git uses -1 by default.

Also allow -1, which python's zlib module accepts, and might
be configured per the git config documentation.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>