repo: honour repository's pack.packSizeLimit
Today, bup honours, in the following order:
 * the command line, and
 * the local repo's pack.packSizeLimit.

This is a bit strange, since this value typically would
be something you'd want to set per repository.

Now that we have remote config handling, honour it when
available, so the order becomes:

 * the command line,
 * the storage repo's pack.packSizeLimit, and
 * the local repo's pack.packSizeLimit.

Note that the storage repo and the local repo might of course
be the same (in case the backup isn't made to a remote repo.)

Again, document this.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>