save: move more tree handling into the stack
"libify" the code a bit here, and move it into the stack classes
that keep track instead of open-coding the list in the save code.
That makes the code easier to handle and simplifies tree-splitting
later, since that requires access to the combined data/metadata
for the items at a given level, which we now get very easily.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
---
v2:
 * use new stack design that keeps a stack, not chained parents,
   and the 'stack' variable again refers to the entire stack
   (though in a different way than before), not just the current
   level