vint: implement the typical pack() in C
This estimates the size, and if we run over it falls back
to the slower python, and similarly if anything happens
with the vuints/vints contained in there, but it still
speeds up the typical case significantly, and improves
raw index speed (on cached filesystem data) by about 14%
(~19.7k paths/sec to ~22.5k paths/sec).

Since it doesn't handle corner cases (very long strings
or very large numbers) add a few tests that cover the
fallback to python.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>