repo/config: provide a way to write repo config
In order to store certain values that are per repo in the
repo (especially in the future with cloud/encryption), add
a way to write values to the repo configuration.

Also add a "config" sub-command to bup to allow manipulating
the configuration values, for remote usage.

To avoid writing values that make the repo unusable, e.g.
writing via remote when you have no manual way of editing
the config file, check the validity of values that we use
in bup itself where otherwise it'd no longer be possible
to use the repo.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>