vfs: unify tree_items/_with_metadata
It's a bit weird that tree_items_with_metadata calls tree_items,
but tree_items is also called directly from the top-level call.
It's also messy to then add support for tree-split.

Unify this by renaming tree_items to _tree_items, marking it as
a private helper function, and tree_items_with_metadata to just
tree_items, while giving it a new "metadata" argument.

Also move the handling of the '.' entry out, this unifies it to
a single location.

This cleans up the code a bit, and in particular simplifies the
tree-split code a lot.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>