vfs: handle latest/ requests specially
If we call rev_list for when we really just need the latest,
we may spend a lot of time doing that. Make this case special
and just read the commit object directly.

Note that we don't create a FakeLink() in this case as that
would result in doing another rev_list() for the link target,
which we wanted to avoid ... And we cannot cache, because the
cache is for the whole rev-list.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>