server: allow forcing a directory, use for "bup on"
When a directory is forced, the client cannot arbitrarily
select any directory, it's always forced into this one. This
can be used with force SSH commands, i.e. if

  command="bup --bup-dir /some/where/ server --force-repo"

is specified for a given authorized SSH key, the client will
not be able to write outside of "/some/where/".

Also use it when we do "bup on" since the reverse:// protocol
doesn't normally init/change the repo, but a malicious remote
could actually attempt to do so. Now, with this, it cannot
escape the bup dir we wanted to use, i.e.

  bup -d /some/where/ on <remote> ...

again will not write outside of "/some/where/", regardless of
what the remote tries to do (unmodified bup will not try to
change it at all.)

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>