server: add and use --mode option
To restrict what a remote bup can do on a server, e.g. in
'bup on' scenarios, add a --mode argument to the server,
adding the following modes:
  - read,
  - append,
  - read-append and
  - unrestricted.

To make this reasonably safe in the face of future additions,
this employs an allowlist of permitted commands, and is based
on a filter put into the base protocol class.

For 'bup on', even restrict when a server is started/connected,
and only do that if the remote command is explicitly listed as
requiring append, read, or both (unrestricted is not used used
right now as remote 'gc', 'rm' or 'prune' etc. aren't supported
at all over the server).

Right now, append is almost equivalent to unrestricted, except
it doesn't allow making ref updates that would drop commits or
writing to the config file.

Note that the mode 'read-append' is not used by 'bup on', it
may however be useful in conjunction with SSH forced commands.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>