web: prepare for more query arguments
Introduce a QueryArgs class that keeps the values that we
care about from the query string, currently that's only
the value of 'hidden', i.e. whether hidden files are shown
or not.

But having this class that can create the needed string
(or bytes) and be used to pass around all the information
makes it much easier to add more arguments.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>