web: generate a proper error if file reading fails
Due to the async nature of this, if an error occurs
we just lose ourselves in exceptions and close the
connection, leaving the browser without any idea of
what might have happened.

Defer header setting until the very last moment
(just before the first write) and if any error has
happened until the first write we can then set the
status properly.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>