PackWriter: match git's pack names
As reported by Jamie Wyrick, git appears to use the sha1 of
the entire pack file as the (default) name for it, not the
sha1(sorted-object-list) as the git-index-pack man page seems
to imply, since it says:

  Once the index has been created, the list of object names is
  sorted and the SHA-1 hash of that list is printed to stdout.
  If --stdin was also used then this is prefixed by either
  "pack\t", or "keep\t" if a new .keep file was successfully
  created. This is useful to remove a .keep file used as a lock
  to prevent the race with git repack mentioned above.

while also saying:

  If <pack-file> is not specified, the pack is written to
  objects/pack/ directory of the current Git repository with a
  default name determined from the pack content.

Originally, git-index-pack was used by bup, and when that was
changed, the naming convention was kept, presumably according
to this documentation; see commit 4cab9ab71fff ("Write idxs
directly rather than using git-index-pack.")

Change things to use the pack's entire content sha1, that we
calculate anyway, as the name.

This requires updating the 'gc' test to not compare by name
but by (index) content.

Also add a test that our behaviour matches git's.

Reported-by: Jamie Wyrick <terrifiedquack80@gmail.com>
Signed-off-by: Johannes Berg <johannes@sipsolutions.net>