bup: add a test for index contents
Add a test that checks that our index files are exactly
identical to the ones git would write. This is slightly
dangerous as git might introduce a new format, but that
seems unlikely now, it's been stable for a long time;
if it happens we can deal with it then.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>