helpers: add a ProgressBar helper class
This can be used like such:

    with ProgressBar(1000) as p:
        for i in range(1001):
            time.sleep(0.023)
            p.update(i, "Receiving %d/%d." % (i, 1000))

and will estimate the remaining time etc.

Or alternatively, with None as the total, and then it's
just a moving bar indicating progress is being made.

For git calling midx/bloom, suspend the current progress
bar via the new helpers.suspend_progress() context manager.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>