repo: make VFS cache ID a repo instance member
This clarifies the name, rather than the current "id"
which can be misunderstood easily as the builtin id().
Since the VFS is the only user and it's only for the
in-memory cache, not for something like the repo-id
(to be added later), just call it vfs_cache_id. Make
it a pure instance member so that no method calls is
needed for it.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>