repo: add a "repo-id" and use it for the index-cache
If you have multiple repos on the same machine, or your host
name changes, or such a thing then the index-cache always gets
very confused and can end up downloading everything over and
over again.

Introduce a concept of a "repo-id" which we store in the config
file under the bup.repo-id key, and generate one automatically
when a repository is opened.

Then, using the existing config machinery, transfer that from
the remote repository, to use it for the index-cache location.

We may find other uses for it as well in the future.

In the VFS test, need to initialize the local test repo after
'bup init' has been called, otherwise it crashes. No need for
it before that point anyway.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>