client: when we have a repo id, use it for the index-cache name
If you have multiple repos on the same machine, or your host
name changes, or such a thing then the index-cache always gets
very confused and can end up downloading everything over and
over again.

So whenever we can get a bup.repo.id from the remote, use that,
prefixed with "id--" (so it can't collide with the existing names
because they can't have a dash) for the index-cache directory name
instead.

Also try to move the cache from the old location to the new one if the
old one exists, so the upgrade path doesn't re-download it all.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: rebase; handle servers without config-get]
[rlb@defaultvalue.org: add id-- prefix]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>