Provide the default index location via path.default_fsindex
Just like with the cache, there's no reason for the bupindex to live
in a real repository. As a consequence, move the path-finding to
path.default_fsindex() and remove the need for a full bup repository
just to run index, or even (remote) save.

To test that, move the 'bup init' invocation far down in the index
tests, to just before we actually save anything.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>