path/index: move index out of git.repo()
Just like with the cache, there's no reason for the bupindex
to live in a real repository. As a consequence, move out the
path-finding to it to path.index() and remove the need for a
full bup repository just to run index, or even (remote) save.

To test that, move the 'bup init' invocation far down in the
index tests, to just before we actually save anything.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>