commands: move/remove various git.check_repo_or_die() calls
If we use LocalRepo() or repo.from_opts() or such then there's
no need to call this explicitly - and in some cases such as in
'bup ls' it stops us from doing work without a local repo even
if we don't need one (such as 'bup ls -r ...').

Remove some calls where it's obviously safe to do so, however,
put a call into LocalRepo() so we still preserve the same error
behaviour.

In bloom/midx, make the call only if they're called to operate
on the repo rather than an explicitly specified directory.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>