save: fix symlink target race
If the symlink target changes while save is running, we can end up
reading one target and recording it in the metadata, but then
recording a different one (with perhaps a different size) in the
blob.

Fix this by not reading the symlink again, but just using the one
we already have in the metadata.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>