save: skip/balk if an entry changed type since indexing
If an entry changed file type (link, regular file, ...) since
indexing, then all kinds of weird things can happen. Skip the
item in such cases and record an error.

This also requires adjusting the test that actually provokes
a failure in metadata read - now we don't store the file at
all, so modify the test to account for that behaviour in the
check.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>