test: add a further metadata confusion test
Since we don't update the metadata in the index if it changed,
setting IX_HASHVALID is somewhat wrong: if we've stored metadata
differently in the repo than what's in the index, then we don't
know that again later if IX_HASHVALID is true, and store wrong
metadata, e.g. wrong size, as demonstrated by this test.

This can also happen with symlinks (target changing between index
and save) and other scenarios, in this case we'd record different
targets as the blob and metadata.

I don't know of a way to fix this right now, we can't update the
index during save to have the correct data since we can't rewrite
it, only update some fixed-size fields - but the size and symlink
target are variable size.

Perhaps we could introduce IX_METAVALID and not set that in such
cases, and then the next save still has to look at it?

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>