aws: cache file size
Hitting dynamodb every time we need to read something from disk
is very expensive - cache the file size in the ranges file (with
"#size=" prefix to find it and to be backward compatible). This
drastically reduces the amount of time needed to do 'bup ls' or
such if the data is already cached.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>