bup: make demux errors clearer
If we hit the

    assert(n <= MAX_PACKET)

in the demux code then quite likely we've been reading
something that shouldn't even be demuxed, but got some
error message from ssh instead of a muxed connection
record.

Clarify this case a bit by printing out the data that
we got (under the assumption it's an error message) and
raising Exception("Connection broken") instead of just
asserting.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>