test-get: split replace test
The replace test is basically the critical path when a lot of
parallelization is available, starting at the beginning and
ending about last. Split it up further to change that.

For me, this reduces test time from ~01:27 to ~1:22, so I guess
we need to do it for the other get tests too...

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>