bup-config: document index impact of bup.split.trees
Since bup.split-trees causes split trees to have a different
sha1 than unsplit trees, and these values are recorded in the
index for later skipping if they exist in the repository, any
use of the same index for multiple repositories with different
values for bup.split-trees will lead to suboptimal behaviour.
Document this.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
---
v2: text editing/clarification, thanks to Greg for the review