Add simple bup-select-fields for platforms without cut -z
This will allow us to output multiple fields (e.g. when the
computation of each is expensive and can be done in the same pass,
cf. check-refs) when we also need \0 (e.g. for unescaped
paths/users/groups) and still provide a way to extract them correctly,
even on platforms that don't support cut -z.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>