git: walk_object: fix item order
There's a (potential) issue with 'bup get' in how it transfers objects
due the the iteration order that we get from walk_object. Consider a
tree fragment like this:

  ... ---- tree1 +---- blob1a
                 +---- blob1b

Once it arrives at tree1, walk_object() will yield the objects in this
order:
 - tree1
 - blob1a
 - blob1b

If, due to prior objects transfered, 'blob1a' completes a pack file in
the destination repository, that means the objects 'tree1' and 'blob1a'
are permanently stored in the destination repository. However, if bup
get is now somehow interrupted before the next file is completed, the
destination repo still has 'tree1' and 'blob1a', although at this point
they're not connected to anything.

However, a further bup get that wants to transfer 'tree1', perhaps just
the same command called again since the previous one was aborted, will
now find that 'tree1' already exists and thus not recurse into it. Once
the second transfer completes, the destination repo is broken, it has
'tree1' connected somewhere, but is missing 'blob1b'.

To fix this issue, change the tree walk order in walk_object() which is
used by bup get so that we walk "leaves first", i.e.:
 - blob1a
 - blob1b
 - tree1

(though the order of blobs doesn't matter.)

This way, any interruptions anywhere will result in unconnected data as
before, but will not result in tree objects being written with missing
leaves. Resuming the interrupted bup get command will now work fine.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>