Prefer bup/remote in XDG cache dir for the index cache
It doesn't make sense, and IMHO never really did, to conflate
 * the local default repository
 * index-cache
 * bupindex
all in ~/.bup/. Perhaps it made sense in a world where every
application had everything in ~/.<name>/, but even then it should've
been ~/.bup/{repo,index-cache,bupindex}, not the directory being a
repo by itself already.

For things that rely on cached() (e.g. the client index-cache) prefer
the XDG cache location ($XDG_CACHE_HOME/bup/ or ~/.cache/bup) when the
item in question already exists there, otherwise the default repo if
it exists there, or finally the XDG location if it doesn't exist in
either one. The bupindex will be handled later.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: set use contextmanagers in common_test_environment]
[rlb@defaultvalue.org: set XDG_CACHE_HOME in test/ext/conftest.py]
[rlb@defaultvalue.org: add release note and bup-config(5) info]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>