repo: support retrieving data for some object types
For walk_object(include_data=False) we need to only retrieve
data for commit/tree objects, but not blobs. Change the get()
API to allow include_data to be a tuple, and implement the
corresponding logic for that depending on the repo type.

Use it in walk_object() to not retrieve all the blob data in
GC's mark phase.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>