vfs: tree_items: use a list for tree items (split trees)
Since tree_decode() is an interator, the commit below introduced a bug
since we iterate twice, and need to iterate from the beginning
twice. Since for ordering reasons the code will load the whole tree
into memory anyway, simply convert it to a list immediately.

This is a second version of the previous commit 269d73a95efe
("vfs: tree_items: use a list for tree items") which fixed this
for unsplit trees, here for split trees.

Fixes: 02ef10bd4366 ("vfs.tree_items: only parse the tree once")
Signed-off-by: Johannes Berg <johannes@sipsolutions.net>