vfs: tree_items: use a list for tree items
Since tree_decode() is an interator, the commit below
introduced a bug since we iterate twice, and need to
iterate from the beginning twice. Since for ordering
reasons the code will load the whole tree into memory
anyway, simply convert it to a list immediately.

Add a test that has a file before the .bupm entry, so
that we test for this; only files before .bupm were
consumed by the first iteration, so the bug didn't get
hit in the tests.

Fixes: 02ef10bd4366 ("vfs.tree_items: only parse the tree once")
Signed-off-by: Johannes Berg <johannes@sipsolutions.net>