Drop stdin/stderr output filtering
This started with "bup newliner" which interposed itself in a pipeline
with subcommands running in a separate process. Then to address
various issues with subprocess management there (signals, etc.)  we
replaced the newliner pipeline with an in-process filter thread. We
also moved module-based commands into the top-level bup process.

This approach has had it's own issues (see the preceeding commit), and
is a lot of complexity (see this removal) for benefits we may be able
to provide in other ways. So for now, remove all of it with the
intention to address any resulting concerns as they arise.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>