LocalRepo: fix failures partway through initialization
Previously, if anything in __init__ raised an exception, closed would
be False, and __del__ could fail as a result. The close and so
__exit__ methods also rely indirectly on _packwriter which is only
initialized after closed is set to False, but right now nothing in
between should throw, so _packwriter should still be available before
finish_writing tries to use it.

Ensure __init__ always leaves the instance in an acceptable state by
setting closed to True at the top and False at the bottom, which
appears to be sufficient with the current process.

This problem was observed during "make -j5 check" via a failure of the
self.closed assertion in __del__.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>