progress: clear line to length of previous message when appropriate
When the previous progress lines ends in a carriage return and is
longer than the current line, print enough spaces to clear the
previous line before printing the new message.

Call progress('') in main after every python subcommand to make sure
we clear the line for the next command. Previously, if your shell
prompt (PS1) was short enough, it would intermingle with bup's last
progress line for commands like validate-object-links.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>