vfs: don't replace (dir) metadata size None with 0
If we don't know the size, don't pretend we do know
it's 0. This likely really just affects the display
in ls/web for directories right now, however, it'll
start mattering more for 'bup rewrite' which would
otherwise always fill in "0" for directories that
should have "None" for their size (i.e. not stored.)

Update the tests to check against None now and to
compare dir size to "-" rather than "0" in ls.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>