Fix CFLAGS and add -fwrapv -fno-strict-aliasing, etc. when available
I thought we were using -fwrapv and -fno-strict-aliasing, but if so,
maybe they were being provided by python-config and now aren't. Add
them explicitly, and a few others that might be helpful. We can back
off if any cause trouble.

Doing this revealed bugs with the handling of bup_config_cflags. Fix
them.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>