HashSplitter_read: read full blocks (retry short reads)
Any time we read, either via read(2) or via some python object's
read() method, we must expect a short read. This will especially
happen on pipes (stdin for bup split) on Linux, which typically return
64k in one read.

If the data read is smaller than the maximum blob size, the hash
splitter will do a lot of extra work, trying to find split points in
the growing data over and over again from the start, all while
re-allocating the buffer and copying the data around.

Handle short reads in the read loop, and read to fill the buffer (or
until EOF.)

Fixes: 4ec5c153c412 ("hashsplit: implement hashsplitting in C and make bits configurable")
Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>