Replace objcache_maker with allow_duplicates/deduplicate_writes
Given the other changes, it's clearer that objcache_maker is only
needed in order to configure deduplication, so simplify by replacing
it with that.

Use the allow_duplicates inverse of deduplicate_writes for
function/method argument names so that it's possible to use None as
the "unset" indicator without the potential for conflation with None
as false that we'd have with a deduplicate_writes=None default.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>