vfs: fix tree split level marker handling
Since the tree split level marker is identified by the .bupd
extension, just like any bup specific entry in tree objects, don't
check for the .bupd. prefix but rather for the .bupd suffix.

Also add such a file to the sampledata, which if saved and then looked
at via the vfs will cause an exception without this fix, e.g. in
test-meta.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>