Add find_tree_entry to find bupm files more efficiently
Though this is small, it may not be worth the complexity.  But we do
need to find the bupm independently of any other parsing of the same
tree fairly often. That's because the metadata for a tree is in the
.bupm inside it, so something like "ls -la
/some/latest/contains-large-dirs/" (or the equivalent fuse/web
operations) will need the .bupm inside every large child, but none of
the rest of the entries.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>